Fix Trouble #4

+++ src/figurecheck.h
@@ -1,4 +1,4 @@
 #ifndef FIGURECHECK_H
 #define FIGURECHECK_H
-void inpsh(char s[], int &x, int &x1, int &x2, int &x3, int &y, int &y1, int &y2, int &y3, float &r);
+void inpsh(char s[], int &x, int &x1, int &x2, int &x3, int &y, int &y1, int &y2, int &y3, int &r);
 #endif
